pyproject.toml
[tool.poetry]
name = "abcde"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
fastapi = {extras = ["standard"], version = "^0.124.4"}
sqlmodel = "^0.0.27"
alembic = "^1.17.2"
psycopg2-binary = "^2.9.11"
sqlalchemy-utils = "^0.42.0"
httpx = "^0.28.1"
requests = "^2.32.5"
numpy = "^2.4.1"
pydantic-settings = "^2.12.0"
python-dateutil = "^2.9.0.post0"
sqlalchemy = "^2.0.46"
aiosqlite = "^0.22.1"

[tool.poetry.group.dev.dependencies]
pytest = "^9.0.2"
pytest-cov = "^7.0.0"
pytest-mock = "^3.15.1"
ruff = "^0.14.9"
pylint = "^4.0.4"
poetry-plugin-export = "^1.10.0"
pylint-pydantic = "^0.4.1"

[tool.poetry.scripts]
docs = "abcde.main:generate_docs"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
include = ["*.py"]
line-length = 88
indent-width = 4
fix = true
target-version = "py310"

[tool.pytest.ini_options]
addopts = "--cov=abcde --cov-report=term-missing"
testpaths = ["tests"]

[tool.coverage.report]
fail_under = 90
show_missing = true
skip_covered = true

[tool.pylint.messages_control]
disable = [
    "missing-module-docstring",
    "missing-class-docstring",
    "missing-function-docstring",
    "fixme"
]
